requirements.txt
mysql-connector-python
schedule
pyyaml
